Arc2Face/requirements_controlnet.txt
numpy<1.24.0
diffusers==0.23.0
transformers==4.34.1
peft
accelerate
insightface
onnxruntime-gpu
gradio
face-alignment
omegaconf
pytorch-lightning==1.4.9
torchmetrics==0.6.2
torchfile
facenet-pytorch
chumpy